Updated PhysicsManager. Find out that terrain physics object with small grid squares marginally reduces performance. Adjusted ObjParser to always load objects scaled to max radius of 1.

+++ ghost/TextureAtlas.cpp
@@ -150,7 +150,7 @@ bool TextureAtlas::findSpace(TextureHeader& th) {
 		if (th.width > sl.width || th.height + sl.row > height) {
 			continue;
 		}
-		fitIndex = i;
+		fitIndex = (INT32) i;
 		break;
 	}
 	if (fitIndex == -1) {
